requirements.txt
streamlit
pillow
openai
qrcode[pil]
pyperclip
requests
python-dotenv
azure-storage-blob
